refactor: use modules for a15/16 apps
move late loaded metadata into directory
create module for a15/16, routes for a17
use loadChildren for all late loaded routes
