+++ projects/ngx-meta/e2e/a15/src/app/app.module.ts
@@ -18,7 +18,6 @@ import {
 } from '@davidlj95/ngx-meta/open-graph'
 import { NgxMetaTwitterCardModule } from '@davidlj95/ngx-meta/twitter-card'
 import { NgxMetaJsonLdModule } from '@davidlj95/ngx-meta/json-ld'
-import { MetaLateLoadedComponent } from './meta-late-loaded/meta-late-loaded.component'
 
 @NgModule({
   declarations: [
@@ -26,13 +25,13 @@ import { MetaLateLoadedComponent } from './meta-late-loaded/meta-late-loaded.com
     MetaSetByServiceComponent,
     MetaSetByRouteComponent,
     MetaSetByRouteAndServiceComponent,
-    MetaLateLoadedComponent,
   ],
   imports: [
     BrowserModule,
     AppRoutingModule,
     NgForOf,
     RouterOutlet,
+    JsonPipe,
     NgxMetaCoreModule.forRoot({ defaults: DEFAULTS }),
     NgxMetaRoutingModule.forRoot(),
     NgxMetaStandardModule,
@@ -40,8 +39,6 @@ import { MetaLateLoadedComponent } from './meta-late-loaded/meta-late-loaded.com
     NgxMetaOpenGraphProfileModule,
     NgxMetaTwitterCardModule,
     NgxMetaJsonLdModule,
-    JsonPipe,
-    JsonPipe,
   ],
   providers: [],
   bootstrap: [AppComponent],
