+++ projects/ngx-meta/e2e/a15/src/app/meta-late-loaded/meta-late-loaded.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core'
+import { RouterModule } from '@angular/router'
+import {
+  LATE_LOADED_METADATA_JSON,
+  provideLateLoadedMetadata,
+} from './late-loaded-metadata'
+import { NgxMetaMetadataLoaderModule } from '@davidlj95/ngx-meta/core'
+import { MetaLateLoadedComponent } from './meta-late-loaded.component'
+import { JsonPipe } from '@angular/common'
+
+@NgModule({
+  declarations: [MetaLateLoadedComponent],
+  imports: [
+    JsonPipe,
+    NgxMetaMetadataLoaderModule,
+    RouterModule.forChild([
+      {
+        path: '',
+        component: MetaLateLoadedComponent,
+        data: { meta: LATE_LOADED_METADATA_JSON },
+      },
+    ]),
+  ],
+  providers: [provideLateLoadedMetadata()],
+})
+export class MetaLateLoadedModule {}
