+++ projects/ngx-meta/e2e/a17/src/app/app.routes.ts
@@ -5,12 +5,7 @@ import { MetaSetByRouteComponent } from './meta-set-by-route/meta-set-by-route.c
 import { MetadataRouteData } from '@davidlj95/ngx-meta/routing'
 import METADATA_JSON from '../../../cypress/fixtures/metadata.json'
 import { MetaSetByRouteAndServiceComponent } from './meta-set-by-route-and-service/meta-set-by-route-and-service.component'
-import { MetaLateLoaded } from './meta-late-loaded/meta-late-loaded.component'
-import {
-  LATE_LOADED_METADATA_JSON,
-  provideLateLoadedMetadata,
-} from '../late-loaded-metadata'
-import { provideNgxMetaMetadataLoader } from '@davidlj95/ngx-meta/core'
+import { META_LATE_LOADED_ROUTES } from './meta-late-loaded/meta-late-loaded.routes'
 
 const metadataRouteData: MetadataRouteData = { meta: METADATA_JSON }
 
@@ -31,8 +26,6 @@ export const routes: Routes = [
   },
   {
     path: ROUTES.metaLateLoaded.path,
-    component: MetaLateLoaded,
-    data: { meta: LATE_LOADED_METADATA_JSON },
-    providers: [provideLateLoadedMetadata(), provideNgxMetaMetadataLoader()],
+    loadChildren: () => META_LATE_LOADED_ROUTES,
   },
 ]
